Using URL as the key in the dictionary
For some reason, using URLRequest as a key in the dictionary was always
failing (elements could not be found). This is surprising, since, while
originalRequest is a deepCopy of the url, URLRequest is Equatable, so I
am not sure why it used to work before and now stopped working. Using
the URL directly fixes the problem (at least as a temporary solution).