Improve Arabic search
This patch improves Arabic search by fuzzily replacing certain
characters with a set of other acceptable ones (ex ا should match أ, إ,
etc). This works by switching the Arabic SQL query with a like query,
replacing these characters with an _. It then postprocesses them,
filtering out the ones that match the pattern with a character that is
unrelated to the one that was searched with.